added some structure and design to this v2
