+++ src/Components/About.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class About extends Component {
+    render() {
+        return (
+            <div>
+                About
+            </div>
+        );
+    }
+}
+
+export default About;
