+++ src/Components/Header.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+class Header extends Component {
+    render() {
+        return (
+            <h1>
+                Header
+            </h1>
+            // Bring some beautiful animation | picture there
+        );
+    }
+}
+
+export default Header;
