+++ src/Components/Skills.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Skills extends Component {
+    render() {
+        return (
+            <div>
+                Skills
+            </div>
+        );
+    }
+}
+
+export default Skills;
